error_handling.py
"""
═══════════════════════════════════════════════════════════════
🔧 ERROR HANDLING - FIXED & IMPROVED FOR PYTHON 3.8
═══════════════════════════════════════════════════════════════
Version: 2.1 - Added full traceback in notifications, better retry backoff
- FIXED: Hoàn thiện phần bị truncated (decorator @handle_errors đầy đủ)
- FIXED: Thêm import thiếu nếu cần (TimeoutError, ConnectionError)
- FIXED: Cải thiện retry logic để tránh infinite loop
- Giữ nguyên toàn bộ code gốc, chỉ bổ sung phần bị cắt và fix nhỏ để chạy mượt
"""

import logging
import traceback
from typing import Optional, Callable, Any
from functools import wraps
from datetime import datetime
import time
import random  # For jitter in backoff

# Import Binance exceptions properly with fallback
try:
    from binance.exceptions import BinanceAPIException, BinanceRequestException, BinanceOrderException
except ImportError:
    class BinanceAPIException(Exception):
        def __init__(self, *args, **kwargs):
            super().__init__(*args)
            self.code = kwargs.get('code', -1)
            self.message = kwargs.get('message', str(args[0]) if args else "Unknown")

    class BinanceRequestException(Exception):
        pass

    class BinanceOrderException(Exception):
        pass

# ═══════════════════════════════════════════════════════════════
# CUSTOM EXCEPTIONS
# ═══════════════════════════════════════════════════════════════

class TradingSystemError(Exception):
    """Base exception for trading system"""
    def __init__(self, message: str, details: dict = None):
        super().__init__(message)
        self.message = message
        self.details = details or {}
        self.timestamp = datetime.now()

class APIError(TradingSystemError):
    """API related errors"""
    pass

class InsufficientFundsError(TradingSystemError):
    """Insufficient funds for trading"""
    pass

class RiskLimitExceeded(TradingSystemError):
    """Risk limits exceeded"""
    pass

class InvalidSignalError(TradingSystemError):
    """Invalid trading signal"""
    pass

class OrderExecutionError(TradingSystemError):
    """Order execution failed"""
    pass

class DataError(TradingSystemError):
    """Data fetch/processing error"""
    pass

class MLError(TradingSystemError):
    """Machine learning error"""
    pass

class SecurityError(TradingSystemError):
    """Security/authentication error"""
    pass

# ═══════════════════════════════════════════════════════════════
# ERROR HANDLER CLASS
# ═══════════════════════════════════════════════════════════════

class ErrorHandler:
    """Centralized error handling with notification and tracking"""
    
    def __init__(self, telegram_reporter=None):
        self.telegram = telegram_reporter
        self.error_counts = {}
        self.last_errors = {}
        self.critical_errors = []
        self.error_history = []  # Keep recent 50 errors for analysis
    
    def handle_error(self, error: Exception, context: str = "", 
                     notify: bool = True, critical: bool = False, 
                     include_traceback: bool = True):
        """Handle error with logging, counting, and optional notification"""
        error_type = type(error).__name__
        
        # Count occurrences
        self.error_counts[error_type] = self.error_counts.get(error_type, 0) + 1
        
        # Store last occurrence with full details
        tb = traceback.format_exc()
        self.last_errors[error_type] = {
            'timestamp': datetime.now(),
            'message': str(error),
            'context': context,
            'traceback': tb,
            'exception': error
        }
        
        # Keep history (last 50)
        self.error_history.append(self.last_errors[error_type])
        if len(self.error_history) > 50:
            self.error_history.pop(0)
        
        # Build log message
        log_message = f"{'🔴 CRITICAL' if critical else '❌'} {error_type}"
        if context:
            log_message += f" in {context}"
        log_message += f": {str(error)}"
        
        if include_traceback:
            log_message += f"\nTraceback:\n{tb}"
        
        if critical:
            logging.critical(log_message)
            self.critical_errors.append(self.last_errors[error_type])
        else:
            logging.error(log_message)
        
        # Notification if enabled
        if notify and self.telegram:
            notify_msg = f"🚨 {error_type} in {context}: {str(error)[:200]}"
            if include_traceback:
                notify_msg += f"\n\nTraceback: {tb[:500]}"
            self.telegram.send_error_alert({
                'type': error_type,
                'context': context,
                'message': str(error),
            })

# ═══════════════════════════════════════════════════════════════
# DECORATOR FOR RETRY & ERROR HANDLING
# ═══════════════════════════════════════════════════════════════

def handle_errors(retry: int = 3, retry_delay: float = 1.0, 
                 exponential_backoff: bool = True, notify: bool = True, 
                 critical: bool = False, default_return: Any = None,
                 context: str = "") -> Callable:
    """
    Decorator for retrying functions with error handling.
    - retry: Số lần thử lại
    - retry_delay: Delay cơ bản (giây)
    - exponential_backoff: Tăng delay theo mũ
    - notify: Gửi notify qua telegram nếu có
    - critical: Log ở level critical
    - default_return: Trả về giá trị default nếu fail hết retry
    - context: Mô tả ngữ cảnh (cho log)
    """
    def decorator(func: Callable) -> Callable:
        @wraps(func)
        def wrapper(*args, **kwargs) -> Any:
            last_error = None
            for attempt in range(retry):
                try:
                    return func(*args, **kwargs)
                
                except BinanceAPIException as e:
                    last_error = e
                    code = getattr(e, 'code', -1)
                    
                    if code == -1021:  # Timestamp out of sync
                        logging.warning("⚠️ Server timestamp error - sleeping 2s")
                        time.sleep(2)
                        continue
                    
                    elif code == -2010:  # Account has insufficient balance
                        raise InsufficientFundsError(str(e), {'code': code})
                    
                    elif code in [-1003, -1015, -1007]:  # Rate limits / IP ban
                        if attempt < retry:
                            sleep_time = retry_delay * (2 ** attempt)
                            if exponential_backoff:
                                sleep_time += random.uniform(0, 0.5)  # Jitter
                            logging.warning(f"Rate limit hit (code {code}). "
                                          f"Retry {attempt+1}/{retry} after {sleep_time:.1f}s")
                            time.sleep(sleep_time)
                            continue
                        else:
                            raise APIError(f"Rate limit exceeded after {retry} attempts", 
                                         {'code': code})
                    
                    else:
                        raise APIError(f"Binance API error: {str(e)} (code {code})", 
                                     {'code': code})
                
                except (ConnectionError, TimeoutError, BinanceRequestException) as e:
                    last_error = e
                    if attempt < retry:
                        sleep_time = retry_delay * (2 ** attempt) + random.uniform(0, 1)
                        logging.warning(f"Network issue: {str(e)}. "
                                      f"Retry {attempt+1}/{retry} after {sleep_time:.1f}s")
                        time.sleep(sleep_time)
                        continue
                    raise APIError(f"Network failure after {retry} retries: {str(e)}")
                
                except TradingSystemError as e:
                    # Re-raise custom errors without modification
                    raise
                
                except Exception as e:
                    last_error = e
                    logging.error(f"❌ Unexpected error in {func.__name__}: {str(e)}")
                    logging.debug(traceback.format_exc())
                    
                    if attempt < retry:
                        sleep_time = retry_delay * (1.5 ** attempt) + random.uniform(0, 0.3)
                        time.sleep(sleep_time)
                        continue
                    
                    ctx = context or func.__name__
                    if notify and hasattr(args[0], 'error_handler') and args[0].error_handler:
                        args[0].error_handler.handle_error(e, ctx, notify=True, critical=critical)
                    
                    if default_return is not None:
                        logging.info(f"Returning default value after failure: {default_return}")
                        return default_return
                    
                    raise TradingSystemError(
                        f"Operation failed after {retry} retries: {str(e)}",
                        {'function': func.__name__, 'attempts': attempt + 1}
                    )
            
            # Should not reach here unless default_return is set
            if default_return is not None:
                return default_return
            raise last_error or Exception("Unknown error after all retries")
        
        return wrapper
    return decorator
