requirements.txt
fastapi
uvicorn[standard]
psutil
requests
urllib3
numpy
pandas
scipy
scikit-learn
joblib
matplotlib
seaborn
cryptography
python-binance
python-telegram-bot
